requirements.txt
zai-sdk==1.0.0
anthropic==0.38.0
psycopg2-binary==2.9.10
fastmcp==0.7.1
